B1: extract goldenmatch-score-core; native scorers become shims over it
Move the three string scorers (jaro_winkler/levenshtein/token_sort), the
private token_sort_string preprocessing, and the score_one dispatch out of
native/src/score.rs into a new pyo3-free lib crate goldenmatch-score-core, so
the native ext and the FFI datafusion-udf crate share ONE source of truth
(parity by construction).

- score-core: plain pub fns; score_one kept VERBATIM incl. the id=2 UNSCALED
  fuzz::ratio ([0,1], not *100) that score_field_matrix depends on.
- native/score.rs: score_one re-imported (pure use); the three #[pyfunction]
  scorers are thin delegating SHIMS (a bare use wont satisfy wrap_pyfunction!).
  lib.rs unchanged. token_sort_string fully removed (grep-zero). Dropped the now
  unused rapidfuzz dep from native (owned by score-core).

Validated: ruff/py_compile N/A (Rust-only). cargo check deferred to CI: the
local cargo cannot unpack the rapidfuzz dep (os error 87 on mtime, corrupted
box). Real verifier is the native parity lane.
