+++ packages/rust/extensions/native/Cargo.toml
@@ -28,7 +28,11 @@ blake2 = "0.10"
 # Canonical record fingerprint canonicalizer (sha256 + JSON parsing live here),
 # shared pyo3-free with the pgrx `postgres` crate so both compute the same id.
 goldenmatch-fingerprint-core = { path = "../fingerprint-core" }
-rapidfuzz = "0.5.0"
+# Canonical string scorers (jaro_winkler / levenshtein / token_sort / score_one),
+# shared pyo3-free with the `datafusion-udf` FFI crate so both surfaces score
+# identically by construction. The #[pyfunction] scorers below are thin shims
+# over this crate; score_one is re-imported verbatim.
+goldenmatch-score-core = { path = "../score-core" }
 rayon = "1.12.0"
 # Arrow C Data Interface for zero-copy kernel input (see score.rs
 # score_block_pairs_arrow). `pyarrow` brings the PyArrowType FFI bridge;
